Temporary workaround for issue found with either popen or os.system functions on some systems.

+++ modules/bibformat/lib/bibformat_engine.py
@@ -213,33 +213,43 @@ def call_old_bibformat(recID, format="HD", on_the_fly=False, verbose=0):
             out += """\n<br/><span class="quicknote">
             Formatting record %i on-the-fly with old BibFormat.
             </span><br/>""" % recID
-##         pipe_input, pipe_output, pipe_error = os.popen3(["%s/bibformat" % bindir,
-##                                                          "otype=%s" % format],
-##                                                         'rw')
+
         # Retrieve MARCXML
         # Build it on-the-fly only if 'call_old_bibformat' was called
         # with format=xm and on_the_fly=True
         xm_record = record_get_xml(recID, 'xm',
                                    on_the_fly=(on_the_fly and format == 'xm'))
 
-        (result_code, result_path) = tempfile.mkstemp()
-        command = "( %s/bibformat otype=%s )  > %s" % (bindir, format, result_path)
-        (xm_code, xm_path) = tempfile.mkstemp()
-        xm_file = open(xm_path, "w")
-        xm_file.write(xm_record)
-        xm_file.close()
-        command = command + " <" + xm_path
-        os.system(command)
-        result_file = open(result_path,"r")
-        bibformat_output = result_file.read()
-        result_file.close()
-        os.remove(result_path)
-        os.remove(xm_path)
-##         pipe_input.write(xm_record)
-##         pipe_input.close()
-##         bibformat_output = pipe_output.read()
-##         pipe_output.close()
-##         pipe_error.close()
+        import platform
+        # FIXME
+        # Some problem have been found using either popen or os.system command.
+        # Here is a temporary workaround until the issue is solved.
+        if platform.python_compiler().find('Red Hat') > -1:
+            # use os.system
+            (result_code, result_path) = tempfile.mkstemp()
+            command = "( %s/bibformat otype=%s )  > %s" % (bindir, format, result_path)
+            (xm_code, xm_path) = tempfile.mkstemp()
+            xm_file = open(xm_path, "w")
+            xm_file.write(xm_record)
+            xm_file.close()
+            command = command + " <" + xm_path
+            os.system(command)
+            result_file = open(result_path,"r")
+            bibformat_output = result_file.read()
+            result_file.close()
+            os.remove(result_path)
+            os.remove(xm_path)
+        else:
+            # use popen
+            pipe_input, pipe_output, pipe_error = os.popen3(["%s/bibformat" % bindir,
+                                                             "otype=%s" % format],
+                                                            'rw')
+            pipe_input.write(xm_record)
+            pipe_input.close()
+            bibformat_output = pipe_output.read()
+            pipe_output.close()
+            pipe_error.close()
+            
         if bibformat_output.startswith("<record>"):
             dom = minidom.parseString(bibformat_output)
             for e in dom.getElementsByTagName('subfield'):
